* Upgrade to Tcl Extension Architecture (TEA) 3.7
Document hints on how to build in a directory other than /usr/local and
in general how to cause it to expect stuff to be in the expected lib dir,
etc, and not in the Tcl build dir.

The answer?  You need to specify where the tclConfig.sh file resides, as in:

    ./configure --prefix=/usr/fa --with-tcl=/usr/fa/lib