Move the support routines such as daemonize, pidfile_open, etc, into a separate file (tcllauncher-support.tcl) from the tcllauncher program and make it so they can be picked up by a non-tcllauncher program by using package require Tcllauncher.
Make the tcllauncher-support.tcl file be generated by expanding tcllauncher-support.tcl.in so the package version will always be generated as configured in the configure.in file.

Add a new pidfile support routine, pidfile_mtime, that returns the mtime of the PID file... useful for a guy who can't obtain the pid to see how long the guy who's holding it has been holding it.

Update the documentation to reflect the above.

Version bump to 1.3.
BUGZID: