Re-assert lock on pidfile during pidfile_write
The documented usage of pidfile_open and pidfile_write did not actually work
because the flock on the pidfile would not survive the daemonize call to fork
to a background process.  Now we re-assert the lock on the pidfile when calling
pidfile_write (aborting if the lock cannot be obtained) so that the pidfile can
be reliably used as a lockfile as documented.